audio and encoder separate
